Split negotiator code
The negotiator turned into a god class with many tasks. The new solution
splits the negotiator into negotiation related code and into SQuAT
algorithm related code (SQuATController). In addition, there are new
classes that make it easier to run and configure SQuAT. However, they
have to be extended in the future.